Resume Aware Chatbot/requirements.txt
# Core LangChain - Latest versions
langchain>=0.3.13
langchain-community>=0.3.13
langchain-openai>=0.2.14
langchain-core>=0.3.28

# Vector Store
faiss-cpu>=1.9.0

# Configuration
python-dotenv>=1.0.1
pydantic>=2.10.5
pydantic-settings>=2.7.1

# API Framework - Latest
fastapi>=0.115.6
uvicorn[standard]>=0.34.0

# PDF handling
PyPDF2>=3.0.1
pymupdf>=1.25.2

# Optional: free local embeddings
sentence-transformers>=3.3.1
torch>=2.5.1

# Token counting
tiktoken>=0.8.0

# Email (if using email notifications)
python-multipart>=0.0.20

